Set things up for deaths
Getting some errors but I'm out of brain energy.
Added a subclass to CharacterController2D (CharacterDude) which will
hopefully allow for easier edits of parameters and maybe different
physics between different characters. (This one is pretty standard
though so the class is empty) (Not sure if we'll end up using it, but I
think it's a good idea to use those for consistency and clarity)

You will need to drag a corpse tile into the corpse value in the editor,
and I don't call LevelController.KillPlayer() anywhere yet. I think
that's it.